Renamed socketcan_interface to clearpath_socket_interface
