+++ clearpath_socketcan_interface/include/clearpath_socketcan_interface/dummy.hpp
@@ -23,9 +23,9 @@
 #include <utility>
 #include <memory>
 
-#include "socketcan_interface/interface.hpp"
-#include "socketcan_interface/dispatcher.hpp"
-#include "socketcan_interface/string.hpp"
+#include "clearpath_socketcan_interface/interface.hpp"
+#include "clearpath_socketcan_interface/dispatcher.hpp"
+#include "clearpath_socketcan_interface/string.hpp"
 
 namespace can
 {
