+++ clearpath_socketcan_interface/include/clearpath_socketcan_interface/threading.hpp
@@ -21,7 +21,7 @@
 #include <memory>
 #include <string>
 
-#include "socketcan_interface/interface.hpp"
+#include "clearpath_socketcan_interface/interface.hpp"
 
 namespace can
 {
