+++ clearpath_socketcan_interface/package.xml
@@ -1,17 +1,16 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format2.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="2">
-  <name>socketcan_interface</name>
+  <name>clearpath_socketcan_interface</name>
   <version>0.8.2</version>
-  <description>Generic CAN interface description with helpers for filtering and driver implementation. Further a socketcan implementation based on boost::asio is included.</description>
+  <description>Clearpath's CAN interface description with helpers for filtering and driver implementation. Further a socketcan implementation based on boost::asio is included.</description>
 
-  <maintainer email="mathias.luedtke@ipa.fraunhofer.de">Mathias Lüdtke</maintainer>
+  <maintainer email="lcamero@clearpathrobotics.com">Mathias Lüdtke</maintainer>
 
   <license>LGPLv3</license>
 
-  <url type="website">http://wiki.ros.org/socketcan_interface</url>
-  <url type="repository">https://github.com/ros-industrial/ros_canopen</url>
-  <url type="bugtracker">https://github.com/ros-industrial/ros_canopen/issues</url>
+  <url type="repository">https://github.com/clearpathrobotics/puma_motor_driver</url>
+  <url type="bugtracker">https://github.com/clearpathrobotics/puma_motor_driver/issues</url>
 
   <author email="mathias.luedtke@ipa.fraunhofer.de">Mathias Lüdtke</author>
