+++ clearpath_socketcan_interface/test/test_string.cpp
@@ -17,7 +17,7 @@
 
 #include <string>
 
-#include "socketcan_interface/string.hpp"
+#include "clearpath_socketcan_interface/string.hpp"
 
 TEST(StringTest, stringconversion)
 {
